@henry40408 2024-03-18
main 1cb0beb

```
Running 10s test @ http://localhost:8080
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     5.21ms  328.51us   8.06ms   86.24%
    Req/Sec     0.96k    33.19     1.02k    63.86%
  19377 requests in 10.11s, 1.64GB read
Requests/sec:   1917.56
Transfer/sec:    166.34MB
```

PR fc66b2e

```
Running 10s test @ http://localhost:8080
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     5.01ms  194.05us   9.63ms   86.50%
    Req/Sec     1.00k    12.17     1.04k    63.00%
  19977 requests in 10.00s, 1.69GB read
Requests/sec:   1996.92
Transfer/sec:    173.22MB
```